Add AC charging speed controls for D3, R3, DP3 and DC controls for D3, R3 (#24)
This PR implements controls for setting AC charging speed as requested in #23 for River 3 (tested), Delta 3 (tested on D3+) and Delta Pro 3. The Ecoflow app does limit input values for some devices and sets lower limits, but it does still work (at least on R3 and D3+).

It also includes DC charging controls for D3 and R3 requested in #25

---------

Co-authored-by: GnoX <dev@gnox.dev>