Add DB write period configuration option to all devices (#30)
Simplified code from #13 that does not modify packet processing, it modifies message processing in devices instead. We still decrypt every message, but only writes new state to DB after configured update period and only deserializes message if it needs to.

The logic for updating fields is little bit more complicated as it has to handle the same message type but with different attributes. This is solved by specifying discriminator fields that determine message types. This can however result in updating sooner than the update period if fields are not disjunctive between the same message type, but it won't increase number of writes from previous version.